[CDF-21666] 👷‍♂️ Remove templates subpackage (#613)
* refactor: remove unused constants

* refactor: moved constants

* refactor: moved data classes

* refactor: moved _migration yaml

* refactor; emptying template utils

* refactor: moved last functions to global utils

* tests: updated test
