+++ cdf-tk-dev.py
@@ -29,7 +29,7 @@ import os
 import sys
 from pathlib import Path
 
-from cognite_toolkit._cdf_tk.templates.data_classes import Environment, InitConfigYAML
+from cognite_toolkit._cdf_tk.data_classes import Environment, InitConfigYAML
 
 REPO_ROOT = Path(__file__).resolve().parent
 root_folder = rf"{REPO_ROOT.absolute()}"
