+++ cognite_toolkit/_cdf_tk/commands/build.py
@@ -7,7 +7,7 @@ import shutil
 import sys
 import traceback
 from collections import ChainMap, defaultdict
-from collections.abc import Hashable, Mapping, Sequence
+from collections.abc import Hashable, Iterator, Mapping, Sequence
 from dataclasses import dataclass, field
 from pathlib import Path
 from typing import Any
@@ -21,7 +21,16 @@ from rich.panel import Panel
 
 from cognite_toolkit._cdf_tk._parameters import ParameterSpecSet
 from cognite_toolkit._cdf_tk.commands._base import ToolkitCommand
-from cognite_toolkit._cdf_tk.constants import _RUNNING_IN_BROWSER
+from cognite_toolkit._cdf_tk.constants import (
+    _RUNNING_IN_BROWSER,
+    EXCL_INDEX_SUFFIX,
+    PROC_TMPL_VARS_SUFFIX,
+    ROOT_MODULES,
+)
+from cognite_toolkit._cdf_tk.data_classes import (
+    BuildConfigYAML,
+    SystemYAML,
+)
 from cognite_toolkit._cdf_tk.exceptions import (
     ToolkitDuplicatedModuleError,
     ToolkitFileExistsError,
@@ -37,17 +46,6 @@ from cognite_toolkit._cdf_tk.loaders import (
     Loader,
     ResourceLoader,
 )
-from cognite_toolkit._cdf_tk.templates._constants import EXCL_INDEX_SUFFIX, PROC_TMPL_VARS_SUFFIX, ROOT_MODULES
-from cognite_toolkit._cdf_tk.templates._utils import (
-    iterate_functions,
-    iterate_modules,
-    module_from_path,
-    resource_folder_from_path,
-)
-from cognite_toolkit._cdf_tk.templates.data_classes import (
-    BuildConfigYAML,
-    SystemYAML,
-)
 from cognite_toolkit._cdf_tk.tk_warnings import (
     FileReadWarning,
     HighSeverityWarning,
@@ -60,7 +58,12 @@ from cognite_toolkit._cdf_tk.tk_warnings import (
     WarningList,
 )
 from cognite_toolkit._cdf_tk.tk_warnings.fileread import DuplicatedItemWarning, MissingRequiredIdentifierWarning
-from cognite_toolkit._cdf_tk.utils import calculate_str_or_file_hash
+from cognite_toolkit._cdf_tk.utils import (
+    calculate_str_or_file_hash,
+    iterate_modules,
+    module_from_path,
+    resource_folder_from_path,
+)
 from cognite_toolkit._cdf_tk.validation import (
     validate_data_set_is_set,
     validate_modules_variables,
@@ -366,7 +369,7 @@ class BuildCommand(ToolkitCommand):
 
         for func in functions:
             found = False
-            for function_subdirs in iterate_functions(module_dir):
+            for function_subdirs in self.iterate_functions(module_dir):
                 for function_dir in function_subdirs:
                     if (fn_xid := func.external_id) == function_dir.name:
                         found = True
@@ -552,6 +555,15 @@ class BuildCommand(ToolkitCommand):
                 )
             )
         return loader
+
+    @staticmethod
+    def iterate_functions(module_dir: Path) -> Iterator[list[Path]]:
+        for function_dir in module_dir.glob(f"**/{FunctionLoader.folder_name}"):
+            if not function_dir.is_dir():
+                continue
+            function_directories = [path for path in function_dir.iterdir() if path.is_dir()]
+            if function_directories:
+                yield function_directories
 
 
 @dataclass
