+++ cognite_toolkit/_cdf_tk/commands/run.py
@@ -19,13 +19,10 @@ from rich.table import Table
 
 from cognite_toolkit._cdf_tk.commands.build import BuildCommand
 from cognite_toolkit._cdf_tk.constants import _RUNNING_IN_BROWSER
+from cognite_toolkit._cdf_tk.data_classes import BuildConfigYAML, SystemYAML
 from cognite_toolkit._cdf_tk.exceptions import ToolkitValidationError
 from cognite_toolkit._cdf_tk.loaders import FunctionLoader, FunctionScheduleLoader
-from cognite_toolkit._cdf_tk.templates import (
-    module_from_path,
-)
-from cognite_toolkit._cdf_tk.templates.data_classes import BuildConfigYAML, SystemYAML
-from cognite_toolkit._cdf_tk.utils import CDFToolConfig, get_oneshot_session
+from cognite_toolkit._cdf_tk.utils import CDFToolConfig, get_oneshot_session, module_from_path
 
 from ._base import ToolkitCommand
