+++ cognite_toolkit/_cdf_tk/utils.py
@@ -24,7 +24,7 @@ import tempfile
 import typing
 from abc import abstractmethod
 from collections import UserDict, defaultdict
-from collections.abc import ItemsView, KeysView, Sequence, ValuesView
+from collections.abc import ItemsView, Iterator, KeysView, Sequence, ValuesView
 from contextlib import contextmanager
 from dataclasses import dataclass, field, fields
 from pathlib import Path
@@ -1139,3 +1139,80 @@ def tmp_build_directory() -> typing.Generator[Path, None, None]:
         yield build_dir
     finally:
         shutil.rmtree(build_dir)
+
+
+def flatten_dict(dct: dict[str, Any]) -> dict[tuple[str, ...], Any]:
+    """Flatten a dictionary to a list of tuples with the key path and value."""
+    items: dict[tuple[str, ...], Any] = {}
+    for key, value in dct.items():
+        if isinstance(value, dict):
+            for sub_key, sub_value in flatten_dict(value).items():
+                items[(key, *sub_key)] = sub_value
+        else:
+            items[(key,)] = value
+    return items
+
+
+def iterate_modules(root_dir: Path) -> Iterator[tuple[Path, list[Path]]]:
+    """Iterate over all modules in the project and yield the module directory and all files in the module.
+
+    Args:
+        root_dir (Path): The root directory of the project
+
+    Yields:
+        Iterator[tuple[Path, list[Path]]]: A tuple containing the module directory and a list of all files in the module
+
+    """
+    # local import to avoid circular import
+    from .constants import EXCL_FILES
+    from .loaders import LOADER_BY_FOLDER_NAME
+
+    if not root_dir.exists():
+        return
+    for module_dir in root_dir.iterdir():
+        if not module_dir.is_dir():
+            continue
+        sub_directories = [path for path in module_dir.iterdir() if path.is_dir()]
+        is_any_resource_directories = any(dir.name in LOADER_BY_FOLDER_NAME for dir in sub_directories)
+        if sub_directories and is_any_resource_directories:
+            # Module found
+            yield module_dir, [path for path in module_dir.rglob("*") if path.is_file() and path.name not in EXCL_FILES]
+            # Stop searching for modules in subdirectories
+            continue
+        yield from iterate_modules(module_dir)
+
+
+@overload
+def module_from_path(path: Path, return_resource_folder: Literal[True]) -> tuple[str, str]: ...
+
+
+@overload
+def module_from_path(path: Path, return_resource_folder: Literal[False] = False) -> str: ...
+
+
+def module_from_path(path: Path, return_resource_folder: bool = False) -> str | tuple[str, str]:
+    """Get the module name from a path"""
+    # local import to avoid circular import
+    from .loaders import LOADER_BY_FOLDER_NAME
+
+    if len(path.parts) == 1:
+        raise ValueError("Path is not a module")
+    last_folder = path.parts[1]
+    for part in path.parts[1:]:
+        if part in LOADER_BY_FOLDER_NAME:
+            if return_resource_folder:
+                return last_folder, part
+            return last_folder
+        last_folder = part
+    raise ValueError("Path is not part of a module")
+
+
+def resource_folder_from_path(path: Path) -> str:
+    """Get the resource_folder from a path"""
+    # local import to avoid circular import
+    from .loaders import LOADER_BY_FOLDER_NAME
+
+    for part in path.parts:
+        if part in LOADER_BY_FOLDER_NAME:
+            return part
+    raise ValueError("Path does not contain a resource folder")
