+++ tests/tests_unit/test_cdf_tk/test_loaders.py
@@ -25,6 +25,12 @@ from pytest_regressions.data_regression import DataRegressionFixture
 
 from cognite_toolkit._cdf_tk._parameters import ParameterSet, read_parameters_from_dict
 from cognite_toolkit._cdf_tk.commands import BuildCommand, CleanCommand, DeployCommand
+from cognite_toolkit._cdf_tk.data_classes import (
+    BuildConfigYAML,
+    Environment,
+    InitConfigYAML,
+    SystemYAML,
+)
 from cognite_toolkit._cdf_tk.exceptions import ToolkitYAMLFormatError
 from cognite_toolkit._cdf_tk.loaders import (
     LOADER_BY_FOLDER_NAME,
@@ -53,17 +59,12 @@ from cognite_toolkit._cdf_tk.loaders import (
     ViewLoader,
 )
 from cognite_toolkit._cdf_tk.loaders.data_classes import NodeAPICall, NodeApplyListWithCall, RawDatabaseTable
-from cognite_toolkit._cdf_tk.templates import (
+from cognite_toolkit._cdf_tk.utils import (
+    CDFToolConfig,
     module_from_path,
     resource_folder_from_path,
+    tmp_build_directory,
 )
-from cognite_toolkit._cdf_tk.templates.data_classes import (
-    BuildConfigYAML,
-    Environment,
-    InitConfigYAML,
-    SystemYAML,
-)
-from cognite_toolkit._cdf_tk.utils import CDFToolConfig, tmp_build_directory
 from cognite_toolkit._cdf_tk.validation import validate_resource_yaml
 from tests.constants import REPO_ROOT
 from tests.tests_unit.approval_client import ApprovalCogniteClient
